requirements.txt
python-dotenv
streamlit
langchain
langchain-openai
langchain-core
langchain-community
openai
faiss-cpu